Format and style a table for LaTeX output

This function formats and styles a table for LaTeX output using the kableExtra package.

zkbl(
  table,
  booktabs = TRUE,
  escape = TRUE,
  format.args = list(decimal.mark = ".", big.mark = ","),
  ...
)

Arguments

table

The data table to be formatted.

booktabs

Logical indicating whether to use the booktabs formatting (default: TRUE).

escape

Logical indicating whether to escape special characters in the table (default: TRUE).

...

Additional arguments to be passed to the kbl function.

Value

A LaTeX-formatted table with specified styling.

Examples

zkbl(mtcars[1:5,])
#> <table class="table table-striped table-hover" style="width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;position: sticky; top:0; background-color: #FFFFFF;">   </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> mpg </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> cyl </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> disp </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> hp </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> drat </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> wt </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> qsec </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> vs </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> am </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> gear </th>
#>    <th style="text-align:right;position: sticky; top:0; background-color: #FFFFFF;"> carb </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> Mazda RX4 </td>
#>    <td style="text-align:right;"> 21.0 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 160 </td>
#>    <td style="text-align:right;"> 110 </td>
#>    <td style="text-align:right;"> 3.90 </td>
#>    <td style="text-align:right;"> 2.620 </td>
#>    <td style="text-align:right;"> 16.46 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 4 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Mazda RX4 Wag </td>
#>    <td style="text-align:right;"> 21.0 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 160 </td>
#>    <td style="text-align:right;"> 110 </td>
#>    <td style="text-align:right;"> 3.90 </td>
#>    <td style="text-align:right;"> 2.875 </td>
#>    <td style="text-align:right;"> 17.02 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 4 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Datsun 710 </td>
#>    <td style="text-align:right;"> 22.8 </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 108 </td>
#>    <td style="text-align:right;"> 93 </td>
#>    <td style="text-align:right;"> 3.85 </td>
#>    <td style="text-align:right;"> 2.320 </td>
#>    <td style="text-align:right;"> 18.61 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 1 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Hornet 4 Drive </td>
#>    <td style="text-align:right;"> 21.4 </td>
#>    <td style="text-align:right;"> 6 </td>
#>    <td style="text-align:right;"> 258 </td>
#>    <td style="text-align:right;"> 110 </td>
#>    <td style="text-align:right;"> 3.08 </td>
#>    <td style="text-align:right;"> 3.215 </td>
#>    <td style="text-align:right;"> 19.44 </td>
#>    <td style="text-align:right;"> 1 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 1 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Hornet Sportabout </td>
#>    <td style="text-align:right;"> 18.7 </td>
#>    <td style="text-align:right;"> 8 </td>
#>    <td style="text-align:right;"> 360 </td>
#>    <td style="text-align:right;"> 175 </td>
#>    <td style="text-align:right;"> 3.15 </td>
#>    <td style="text-align:right;"> 3.440 </td>
#>    <td style="text-align:right;"> 17.02 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 0 </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 2 </td>
#>   </tr>
#> </tbody>
#> </table>