Print a Styled Table for HTML or LaTeX Output

This function prints a styled table in either HTML or LaTeX format, depending on the output document type.

print_table(
  df,
  caption = "",
  digits = 2,
  print_rownames = FALSE,
  fontsize = 9,
  big.mark = ",",
  na_str = "",
  ...
)

Arguments

df

The data frame to be formatted as a table.

caption

A character string specifying the caption for the table (default: ”).

digits

An integer indicating the number of digits to display (default: 2).

print_rownames

Logical value indicating whether to print row names (default: FALSE).

fontsize

Font size for the table (default: 9).

big.mark

A character string used as the big mark separator for numbers (default: ',').

na_str

A string used to replace missing values (default: ”).

...

Additional arguments to be passed to the formatting functions.

Value

A styled table in HTML or LaTeX format, depending on the document type.

Examples

head(mtcars) %>% print_table()

mpg

cyl

disp

hp

drat

wt

qsec

vs

am

gear

carb

21.00

6.00

160.00

110.00

3.90

2.62

16.46

0.00

1.00

4.00

4.00

21.00

6.00

160.00

110.00

3.90

2.88

17.02

0.00

1.00

4.00

4.00

22.80

4.00

108.00

93.00

3.85

2.32

18.61

1.00

1.00

4.00

1.00

21.40

6.00

258.00

110.00

3.08

3.21

19.44

1.00

0.00

3.00

1.00

18.70

8.00

360.00

175.00

3.15

3.44

17.02

0.00

0.00

3.00

2.00

18.10

6.00

225.00

105.00

2.76

3.46

20.22

1.00

0.00

3.00

1.00