Format and style a table for LaTeX output

This function formats and styles a table for LaTeX output using the kableExtra package.

zkbl(table, booktabs = TRUE, escape = TRUE, ...)

Arguments

table

The data table to be formatted.

booktabs

Logical indicating whether to use the booktabs formatting (default: TRUE).

escape

Logical indicating whether to escape special characters in the table (default: TRUE).

...

Additional arguments to be passed to the kbl function.

Value

A LaTeX-formatted table with specified styling.

Examples